fix quit share
