+++ lib/Alchemy/Phrasea/Controller/Prod/ShareController.php
@@ -9,11 +9,16 @@
  */
 namespace Alchemy\Phrasea\Controller\Prod;
 
+use Alchemy\Phrasea\Application\Helper\EntityManagerAware;
 use Alchemy\Phrasea\Controller\Controller;
+use Alchemy\Phrasea\Model\Entities\Basket;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 class ShareController extends Controller
 {
+    use EntityManagerAware;
+
     /**
      *  Share a record
      *
@@ -89,5 +94,29 @@ class ShareController extends Controller
         ];
 
         return $this->renderResponse('prod/Share/record.html.twig', $outputVars);
+    }
+
+    public function quitshareAction(Request $request, Basket $basket)
+    {
+        $ret = [
+            'success' => false,
+            'message' => ""
+        ];
+
+        $user = $this->getAuthenticatedUser();
+        if( !is_null($participant = $basket->getParticipant($user))) {
+            $manager = $this->getEntityManager();
+            $manager->beginTransaction();
+
+            $basket->removeParticipant($participant);
+            $manager->remove($participant);
+            $manager->persist($basket);
+            $manager->flush();
+
+            $manager->commit();
+            $ret['success'] = true;
+        }
+
+        return $this->app->json($ret);
     }
 }
