+++ lib/Alchemy/Phrasea/ControllerProvider/Prod/Share.php
@@ -14,6 +14,7 @@ namespace Alchemy\Phrasea\ControllerProvider\Prod;
 use Alchemy\Phrasea\Application as PhraseaApplication;
 use Alchemy\Phrasea\Controller\Prod\ShareController;
 use Alchemy\Phrasea\ControllerProvider\ControllerProviderTrait;
+use Alchemy\Phrasea\Core\LazyLocator;
 use Silex\Application;
 use Silex\ControllerProviderInterface;
 use Silex\ServiceProviderInterface;
@@ -27,6 +28,7 @@ class Share implements ControllerProviderInterface, ServiceProviderInterface
     {
         $app['controller.prod.share'] = $app->share(function (PhraseaApplication $app) {
             return (new ShareController($app))
+                ->setEntityManagerLocator(new LazyLocator($app, 'orm.em'))
                 ;
         });
     }
@@ -44,6 +46,9 @@ class Share implements ControllerProviderInterface, ServiceProviderInterface
         $controllers->before(function () use ($firewall) {
             $firewall->requireNotGuest();
         });
+
+        // tranform 'basket' argument (id) to basket object
+        $controllers->before($app['middleware.basket.converter']);
 
         $controllers->get('/record/{base_id}/{record_id}/', 'controller.prod.share:shareRecord')
             ->before(function (Request $request) use ($app, $firewall) {
@@ -65,6 +70,11 @@ class Share implements ControllerProviderInterface, ServiceProviderInterface
                 }
             })
             ->bind('share_record');
+
+        /** @uses ShareController::quitshareAction() */
+        $controllers->post('/quitshare/{basket}/', 'controller.prod.share:quitshareAction')
+            ->assert('basket', '\d+')
+            ->bind('prod_share_quitshare');
 
         return $controllers;
     }
