+++ lib/Alchemy/Phrasea/Twig/PhraseanetExtension.php
@@ -59,7 +59,7 @@ class PhraseanetExtension extends \Twig_Extension
     {
         return [
             // change this version when you change JS file to force the navigation to reload js file
-            'assetFileVersion' => 71
+            'assetFileVersion' => 72
         ];
 
     }
